pyproject.toml
[project]
name = "vinmeclumina"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.56.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
]
